+++ usr/src/demos/goblet.c
@@ -74,8 +74,6 @@ int main(int argc, char **argv)
     for (f = 0; f < n; f++) {
         scene(f);
         yield();
-        if (argc < 2) sleep(75u);
-        else sleep(1u);
     }
     return 0;
 }